pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clam"
version = "0.1.0"
description = "CLassify Anything Model: Multi-modal classification using embeddings and VLMs for images, audio, and tabular data"
readme = "README.md"
authors = [
    {name = "Ben Feuer", email = "penfever@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "torch",
    "transformers>=4.30.0",
    "datasets>=2.10.0",
    "tabpfn>=0.1.7",
    "scikit-learn>=1.0.0",
    "tqdm>=4.62.0",
    "peft>=0.4.0",
    "openml>=0.13.0",
    "wandb>=0.15.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "accelerate>=0.20.0",
    "jsonschema>=4.0.0",
    "jsonargparse>=4.0.0",
    "PyYAML>=3.13",
]

[project.optional-dependencies]
vision = [
    "torchvision>=0.15.0",
    "Pillow>=8.0.0",
    "timm>=0.9.0",
    # Additional vision dependencies that might be useful
    "opencv-python>=4.5.0",
    "albumentations>=1.3.0",
]

vlm = [
    # Vision Language Model dependencies
    "transformers>=4.45.0",
    "accelerate>=0.26.0",
    # Note: auto-gptq and bitsandbytes require CUDA, skip on Mac
    "optimum>=1.21.0",
]

vlm_cuda = [
    # CUDA-specific VLM dependencies (Linux/Windows with NVIDIA GPU)
    "auto-gptq>=0.7.0",
    "bitsandbytes>=0.43.0",
]

audio = [
    # Audio processing dependencies for audio classification
    "librosa>=0.9.2",
    "soundfile>=0.12.1",
    "openai-whisper>=20231117",
    "torchaudio>=0.13.0",
    "scipy>=1.9.0",
    # CLAP and transformers for audio-text models
    "transformers>=4.35.0",
    "datasets>=2.10.0",
    # Optional audio utilities
    "pydub>=0.25.1",
    "audiomentations>=0.30.0",
]

tabula_8b = [
    "uv>=0.6.0",
    "black==23.11.0",
    "bitsandbytes>=0.45.0",
    "fire>=0.7.0",
    "groovy>=0.1.2",
    "ipdb>=0.13.13",
    "ninja>=1.11.0",
    "optimum>=1.24.0",
    "ruff>=0.11.0",
    "texttable>=1.7.0",
    "torchmetrics>=1.7.0",
    "uvicorn>=0.34.0",
    "fastapi>=0.115.0",
    "gradio>=5.22.0",
    "ray",
    "s3fs>=2025.3.0",
    "safehttpx>=0.1.6",
    "semantic-version>=2.10.0",
    "starlette>=0.46.1",
    "python-multipart>=0.0.20",
    "pydub>=0.25.1",
    "webdataset>=0.2.111",
    "xport>=3.2.1",
    "xxhash>=3.5.0",
    "catboost>=1.2.5",
    "fairlearn",
    "folktables>=0.0.12",
    "ucimlrepo>=0.0.7",
    "boto3==1.34.*",
    "botocore>=1.34.162",
    "jmespath>=1.0.1",
    "s3transfer>=0.9.0",
    "xgboost",
    "nltk",
    # Git dependencies from RTFM repo
    "tableshift @ git+https://github.com/mlfoundations/tableshift.git",
    "llama-recipes @ git+https://github.com/jpgard/llama-recipes.git", 
    "tabliblib @ git+https://github.com/penfever/tabliblib.git",
    # RTFM itself
    "rtfm @ git+https://github.com/penfever/rtfm.git",
]

[project.urls]
Homepage = "https://github.com/penfever/clam"
Repository = "https://github.com/penfever/clam"

[tool.setuptools]
packages = ["clam", "clam.data", "clam.models", "clam.models.vq", "clam.train", "clam.utils"]